Model lifecycle background work as Command values
Replace the free functions passed to performBackgroundTask with a
Command protocol (execute(in:)) and one value type per unit of work
(Trigger/Complete/Recovery), run via NSPersistentContainer.run. Fold
each object's production worker into its command, inject Bundle into the
version/session triggers, and split multi-command files so each nested
type gets its own Type.Command.swift file.

Regroup the Identity cluster into Core/Identity (Identity, Command,
ActionTable, bootstrap and table extensions), move Protected alongside
Identity, rename the former Backend folder to Setup, and rename
MonitorError to LifecycleError.
