+++ Sources/Scout/Core/Lifecycle/Base/Bundle+Version.swift
@@ -1,0 +1,18 @@
+//
+// Copyright 2026 Mikhail Kasianov
+//
+// Use of this source code is governed by an MIT-style
+// license that can be found in the LICENSE file or at
+// https://opensource.org/licenses/MIT.
+
+import Foundation
+
+extension Bundle {
+    var marketingVersion: String? {
+        infoDictionary?["CFBundleShortVersionString"] as? String
+    }
+
+    var buildNumber: String? {
+        infoDictionary?["CFBundleVersion"] as? String
+    }
+}
