+++ Sources/Scout/Core/Lifecycle/Install/InstallObject.Trigger.swift
@@ -1,0 +1,31 @@
+//
+// Copyright 2026 Mikhail Kasianov
+//
+// Use of this source code is governed by an MIT-style
+// license that can be found in the LICENSE file or at
+// https://opensource.org/licenses/MIT.
+
+import CoreData
+
+extension InstallObject {
+    struct Trigger: Command {
+        let installID: UUID
+        let deviceID: UUID
+
+        func execute(in context: NSManagedObjectContext) throws {
+            let request = NSFetchRequest<InstallObject>(entityName: "InstallObject")
+            request.predicate = NSPredicate(format: "installID == %@", installID as CVarArg)
+            request.fetchLimit = 1
+
+            guard try context.fetch(request).isEmpty else {
+                return
+            }
+
+            let install = context.insert(InstallObject.self)
+            install.installID = installID
+            install.date = Date()
+            install.device = try context.existing(DeviceObject.self, key: "deviceID", id: deviceID)
+            try context.save()
+        }
+    }
+}
