+++ Sources/Scout/Core/Setup/SetupError.swift
