+++ Tests/ScoutTests/Core/Lifecycle/Launch/LaunchObjectMonitorTests.swift
@@ -18,7 +18,7 @@ struct LaunchObjectMonitorTests {
 
     @Test("trigger creates a LaunchObject")
     func createsLaunch() throws {
-        try LaunchObject.trigger(launchID: identity.launch, installID: identity.install, in: context)
+        try LaunchObject.Trigger(launchID: identity.launch, installID: identity.install).execute(in: context)
 
         let launches = try context.fetchAll(LaunchObject.self)
         #expect(launches.count == 1)
@@ -28,9 +28,9 @@ struct LaunchObjectMonitorTests {
 
     @Test("Launch stays open when a session inside it is completed")
     func launchRemainsOpenAfterSessionComplete() throws {
-        try LaunchObject.trigger(launchID: identity.launch, installID: identity.install, in: context)
-        try SessionObject.trigger(session: identity.session, launchID: identity.launch, in: context)
-        try SessionObject.complete(launchID: identity.launch, in: context)
+        try LaunchObject.Trigger(launchID: identity.launch, installID: identity.install).execute(in: context)
+        try SessionObject.Trigger(session: identity.session, launchID: identity.launch).execute(in: context)
+        try SessionObject.Complete(launchID: identity.launch).execute(in: context)
 
         let launches = try context.fetchAll(LaunchObject.self)
         #expect(launches.first?.endDate == nil)
