+++ Tests/ScoutTests/Core/Lifecycle/Session/SessionObjectMonitorTests.swift
@@ -19,9 +19,9 @@ struct SessionObjectMonitorTests {
     @Test("complete sets endDate on an open session")
     func completeOpenSession() throws {
         LaunchObject.stub(date: Date(), in: context)
-        try SessionObject.trigger(session: identity.session, launchID: identity.launch, in: context)
+        try SessionObject.Trigger(session: identity.session, launchID: identity.launch).execute(in: context)
 
-        try SessionObject.complete(launchID: identity.launch, in: context)
+        try SessionObject.Complete(launchID: identity.launch).execute(in: context)
 
         let sessions = try context.fetchAll(SessionObject.self)
         #expect(sessions.count == 1)
@@ -30,7 +30,7 @@ struct SessionObjectMonitorTests {
 
     @Test("trigger stamps the session with the current app version")
     func triggerStampsAppVersion() throws {
-        try SessionObject.trigger(session: identity.session, launchID: identity.launch, in: context)
+        try SessionObject.Trigger(session: identity.session, launchID: identity.launch).execute(in: context)
 
         let session = try #require(try context.fetchAll(SessionObject.self).first)
         #expect(session.appVersion == Bundle.main.marketingVersion)
@@ -39,7 +39,7 @@ struct SessionObjectMonitorTests {
 
     @Test("trigger stamps the session with the runtime environment")
     func triggerStampsEnvironment() throws {
-        try SessionObject.trigger(session: identity.session, launchID: identity.launch, in: context)
+        try SessionObject.Trigger(session: identity.session, launchID: identity.launch).execute(in: context)
 
         let session = try #require(try context.fetchAll(SessionObject.self).first)
         #expect(session.osVersion == SystemInfo.osVersion)
@@ -50,12 +50,12 @@ struct SessionObjectMonitorTests {
     @Test("complete is a no-op when the session is already closed")
     func completeTwiceIsNoop() throws {
         LaunchObject.stub(date: Date(), in: context)
-        try SessionObject.trigger(session: identity.session, launchID: identity.launch, in: context)
-        try SessionObject.complete(launchID: identity.launch, in: context)
+        try SessionObject.Trigger(session: identity.session, launchID: identity.launch).execute(in: context)
+        try SessionObject.Complete(launchID: identity.launch).execute(in: context)
 
         let firstEndDate = try #require(try context.fetchAll(SessionObject.self).first?.endDate)
 
-        try SessionObject.complete(launchID: identity.launch, in: context)
+        try SessionObject.Complete(launchID: identity.launch).execute(in: context)
 
         let session = try #require(try context.fetchAll(SessionObject.self).first)
         #expect(session.endDate == firstEndDate)
@@ -63,8 +63,8 @@ struct SessionObjectMonitorTests {
 
     @Test("complete throws notFound when no session exists for the current launch")
     func completeWithoutSessionThrows() throws {
-        #expect(throws: MonitorError.notFound) {
-            try SessionObject.complete(launchID: identity.launch, in: context)
+        #expect(throws: LifecycleError.notFound) {
+            try SessionObject.Complete(launchID: identity.launch).execute(in: context)
         }
     }
 }
